% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_context}
\alias{get_context}
\title{Get Context of Terms in a Document}
\usage{
get_context(.pos, .doc, .context = c("word", "sentence"), .n, .sep = "---")
}
\arguments{
\item{.pos}{A dataframe with terms and their positions in the input .doc dataframe.}

\item{.doc}{A tokenized dataframe of an input text.}

\item{.context}{A character vector specifying the context type, either "word" or "sentence".}

\item{.n}{A positive integer specifying the number of words or sentences to include in the context.}

\item{.sep}{A character string used as a separator between sentences when concatenating tokens.}
}
\value{
A dataframe with pre and post context information for each term.
}
\description{
Get Context of Terms in a Document
}
