% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_context_word}
\alias{get_context_word}
\title{Get Context of Terms in a Document}
\usage{
get_context_word(.pos, .doc, .n)
}
\arguments{
\item{.pos}{A dataframe with terms and their positions in the input .doc dataframe.}

\item{.doc}{A tokenized dataframe of an input text.}

\item{.n}{A positive integer specifying the number of words or sentences to include in the context.}
}
\value{
A dataframe with pre and post context information for each term.
}
\description{
Get Context of Terms in a Document
}
