% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{position_count}
\alias{position_count}
\title{Get Position of Terms}
\usage{
position_count(.tls, .doc, ...)
}
\arguments{
\item{.tls}{A term list prepared by prep_termlist()}

\item{.doc}{A tokenized Dataframe prepared by prep_document()}

\item{...}{Any number of Columns that mark a separator of tokens (e.g. a sentence)}
}
\value{
A Dataframe 5 columns:
doc_id: Document ID of the dataframe in argument .document prepared by prep_document()\cr
hash: Hash value of the terms in argument .termlist prepared by prep_termlist()\cr
start: Starting value (tok_id in .document) of the term\cr
stop: Ending value (tok_id in .document) of the term\cr\cr
dup: Logical indicator if term is contained in higher N-Gram and has already been found
}
\description{
Get Position of Terms
}
