% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{prep_document}
\alias{prep_document}
\title{Prepare Documents}
\usage{
prep_document(.tab, .fun_std = NULL, .until = c("tok", "sen", "par", "pag"))
}
\arguments{
\item{.tab}{A Dataframe with at least two columns:\cr
doc_id: Unique document identifier\cr
text: Document text\cr

IMPORTANT NOTE: Dataframe must contain only one row per doc_id}

\item{.fun_std}{A function to standardize Strings. Default = NULL (no standardization use)}

\item{.until}{c("tok", "sen", "par", "pag")}
}
\value{
A Dataframe with the following columns:\cr
doc_id: Unique document identifier\cr
pag_id: Page ID of token\cr
par_id: Paragraph ID of token\cr
sen_id: Sentence ID of token\cr
tok_id: Token ID\cr
token: Token (tokenized using white spaces)
}
\description{
Prepare Documents
}
\examples{
doc <- prep_document(table_doc, std_str)
}
