% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{std_str}
\alias{std_str}
\title{Basic String Standardization}
\usage{
std_str(.str, .op = c("space", "punct", "case", "ascii"))
}
\arguments{
\item{.str}{A string}

\item{.op}{Any of c("space", "punct", "case", "ascii")}
}
\value{
A string
}
\description{
Basic String Standardization
}
\examples{
std_str("  TesT String")
}
