% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentations.R
\docType{data}
\name{table_doc}
\alias{table_doc}
\title{Example Text}
\format{
A data frame one row and two columns
\describe{
\item{doc_id}{A unique identifier of the document}
\item{text}{Text of the document}
}
}
\usage{
table_doc
}
\description{
A dataset containing an example text from wikipedia
\href{https://en.wikipedia.org/wiki/Natural_language_processing}{Wikipedia}
}
\keyword{datasets}
