% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_documentations.R
\docType{data}
\name{table_terms}
\alias{table_terms}
\title{Example Term List (Short)}
\format{
A data frame 7 rows and 2 columns
\describe{
\item{tid}{A term identifier}
\item{term}{A Term}
\item{part_of_ngram }{Indicator Column if a term is part of a higher ngram}
\item{origin}{Term origin}
}
}
\usage{
table_terms
}
\description{
A dataset containing example terms to be searched within a document
}
\keyword{datasets}
