% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{tl_dep}
\alias{tl_dep}
\title{Helper Function: Get Term Dependencies}
\usage{
tl_dep(.tls)
}
\arguments{
\item{.tls}{A Datframe produced by h_prep_termlist()}
}
\value{
A Dataframe
}
\description{
Helper Function: Get Term Dependencies
}
